+++ src/lexicons/index.ts
@@ -1,4 +1,0 @@
-export * as BlueZioAtfileFinger from "./types/blue/zio/atfile/finger.ts";
-export * as BlueZioAtfileLock from "./types/blue/zio/atfile/lock.ts";
-export * as BlueZioAtfileMeta from "./types/blue/zio/atfile/meta.ts";
-export * as BlueZioAtfileUpload from "./types/blue/zio/atfile/upload.ts";